+++ Form/DataTransformer/RangeToArrayTransformer.php
@@ -14,16 +14,35 @@ use Symfony\Component\Form\Exception\TransformationFailedException;
 class RangeToArrayTransformer implements DataTransformerInterface
 {
     /**
+     * @var string
+     */
+    private $rangeClass;
+
+    /**
+     * @param string $rangeClass
+     * @throws \InvalidArgumentException Throws when given class not a range implementation.
+     */
+    public function __construct($rangeClass = 'NetTeam\DDD\ValueObject\Range')
+    {
+        if ($rangeClass !== 'NetTeam\DDD\ValueObject\Range' && !in_array('NetTeam\DDD\ValueObject\Range', class_parents($rangeClass))) {
+            throw new \InvalidArgumentException('Expected instance of "NetTeam\DDD\ValueObject\Range".');
+        }
+
+        $this->rangeClass = $rangeClass;
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function transform($range)
     {
+
         if (null === $range) {
             return array('min'=> null, 'max' => null);
         }
 
-        if (!$range instanceof Range) {
-            throw new TransformationFailedException('Expected a NetTeam\DDD\ValueObject\Range.');
+        if (!$range instanceof $this->rangeClass) {
+            throw new TransformationFailedException(sprintf('Expected instance of "%s".', $this->rangeClass));
         }
 
         return array('min'=> $range->min(), 'max' => $range->max());
@@ -38,6 +57,6 @@ class RangeToArrayTransformer implements DataTransformerInterface
             throw new TransformationFailedException('Expected an array.');
         }
 
-        return new Range($value['min'], $value['max']);
+        return new $this->rangeClass($value['min'], $value['max']);
     }
 }