+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## CHANGELOG ##
 
 #### v1.1.6 (2014-XX-XX) ####
-- Add support for DateRange in forms.
+- Add support for customized Range classes to RangeType.
 
 #### v1.1.5 (2014-07-03) ####
 - Merge z branchem 1.0